cleaned up python models for serial and mmio
